# src/main/resources/OpenAPI_ViviGo.yaml
openapi: 3.1.0
info:
  title: ViviGo API
  description: API de ViviGo una aplicación web para la gestión de alojamientos (casas, apartamentos, fincas), reservas y comentarios, con roles diferenciados para usuarios y anfitriones, utilizando Spring Boot (JPA) y Angular.
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
components:
  schemas:
...
    Alojamiento:
      type: object
      properties:
        titulo:
          type: string
          minLength: 1
        descripcion:
          type: string
          minLength: 1
        ubicacion:
          $ref: '#/components/schemas/Ubicacion'
        precioNoche:
          type: number
          format: float
          minimum: 0
        capacidad:
          type: integer
          minimum: 1
        servicios:
          type: array
          items:
            type: string
        imagenes:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        imagenPrincipal:
          type: integer
          description: "Índice de la imagen principal en el array de imágenes"
          minimum: 0
      required:
        - titulo
        - descripcion
        - ubicacion
        - precioNoche
        - capacidad
        - imagenes
    Ubicacion:
      type: object
      properties:
        ciudad:
          type: string
          minLength: 1
        direccion:
          type: string
          minLength: 1
        coordenadas:
          $ref: '#/components/schemas/Coordenadas'
      required:
        - ciudad
        - direccion
        - coordenadas
    Coordenadas:
      type: object
      properties:
        latitud:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitud:
          type: number
          format: double
          minimum: -180
          maximum: 180
      required:
        - latitud
        - longitud
...
    PaginatedResponse:
      type: object
      properties:
        contenido:
          type: array
          items: {}
        pagina:
          type: integer
          minimum: 0
        tamano:
          type: integer
          minimum: 1
        totalElementos:
          type: integer
          minimum: 0
        totalPaginas:
          type: integer
          minimum: 0
        esUltima:
          type: boolean
        esPrimera:
          type: boolean
...
    Metricas:
      type: object
      properties:
        numeroReservas:
          type: integer
          minimum: 0
        promedioCalificaciones:
          type: number
          format: float
          minimum: 0
          maximum: 5
        ingresosTotales:
          type: number
          format: float
          minimum: 0
        ocupacion:
          type: number
          format: float
          minimum: 0
          maximum: 100
...
    Error:
      type: object
      properties:
        error:
          type: string
        mensaje:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /alojamientos:
    post:
      summary: Crear un nuevo alojamiento
      tags: [Alojamientos]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Alojamiento'
      responses:
        '201':
          description: Alojamiento creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alojamiento'
        '400':
          description: Faltan datos obligatorios o datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Usuario no es anfitrión
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Buscar alojamientos con filtros
      tags: [ Alojamientos ]
      parameters:
        - name: ciudad
          in: query
          schema:
            type: string
        - name: fecha_entrada
          in: query
          schema:
            type: string
            format: date
        - name: fecha_salida
          in: query
          schema:
            type: string
            format: date
        - name: precio_min
          in: query
          schema:
            type: number
            format: float
            minimum: 0
        - name: precio_max
          in: query
          schema:
            type: number
            format: float
            minimum: 0
        - name: servicios
          in: query
          schema:
            type: array
            items:
              type: string
        - name: capacidad
          in: query
          schema:
            type: integer
            minimum: 1
        - name: pagina
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: tamano
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Lista paginada de alojamientos disponibles (excluye eliminados)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      contenido:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              format: int64
                            titulo:
                              type: string
                            imagenPrincipal:
                              type: string
                              format: uri
                            precioNoche:
                              type: number
                              format: float
                            ubicacion:
                              $ref: '#/components/schemas/Ubicacion'
                            promedioCalificaciones:
                              type: number
                              format: float
                            capacidad:
                              type: integer
        '400':
          description: Parámetros de búsqueda inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: El alojamiento no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /alojamientos/{id_alojamiento}:
    get:
      summary: Ver la información completa de un alojamiento
      tags: [Alojamientos]
      parameters:
        - name: id_alojamiento
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Datos completos del alojamiento con galería, calendario de disponibilidad y comentarios
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Alojamiento'
                  - type: object
                    properties:
                      disponibilidad:
                        type: array
                        items:
                          type: object
                          properties:
                            fecha:
                              type: string
                              format: date
                            disponible:
                              type: boolean
        '404':
          description: Alojamiento no existe o está eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Editar un alojamiento existente
      tags: [Alojamientos]
      security:
        - BearerAuth: []
      parameters:
        - name: id_alojamiento
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                titulo:
                  type: string
                  minLength: 1
                descripcion:
                  type: string
                  minLength: 1
                precioNoche:
                  type: number
                  format: float
                  minimum: 0
                capacidad:
                  type: integer
                  minimum: 1
                servicios:
                  type: array
                  items:
                    type: string
                imagenes:
                  type: array
                  items:
                    type: string
                    format: uri
                  minItems: 1
                  maxItems: 10
                imagenPrincipal:
                  type: integer
                  minimum: 0
                ubicación:
                  $ref: '#/components/schemas/Ubicacion'

      responses:
        '200':
          description: Alojamiento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alojamiento'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos (no es el anfitrión propietario)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Alojamiento no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Eliminar un alojamiento (soft delete)
      tags: [Alojamientos]
      security:
        - BearerAuth: []
      parameters:
        - name: id_alojamiento
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Alojamiento marcado como eliminado
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos (no es el anfitrión propietario)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Alojamiento no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: No se puede eliminar - tiene reservas futuras
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /alojamientos/{id}/metricas:
    get:
      summary: Obtener métricas de alojamiento por rango de fechas
      tags: [ Alojamientos ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: fecha_inicio
          in: query
          schema:
            type: string
            format: date
        - name: fecha_fin
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Métricas del alojamiento en el rango especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metricas'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos (no es el anfitrión propietario)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Alojamiento no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # GESTIÓN DE COMENTARIOS
...